requirements.txt
customtkinter
matplotlib
numpy
open_clip_torch
openpyxl
pandas
Pillow
python-docx
requests
scikit-learn
torch
tqdm
transformers
